pkg: oci: Refactor code and add new signature-verifier package.
Previous code related to gadget verification is now part of its own package in
pkg/signature-verifier.
